backend/requirements.txt
fastapi
uvicorn
supabase
python-multipart 
boto3
pyjwt
python-dotenv
pgvector
geoalchemy2
sqlalchemy
Pillow
requests
opencv-python